feat: API 연결

+++ src/main/java/capsrock/mainPage/dto/Dashboard.java
@@ -3,4 +3,4 @@ package capsrock.mainPage.dto;
 public record Dashboard(
         int maxTemp,
         int minTemp,
-        int temp) {}
+        double temp) {}
